+++ stack/constants.py
@@ -1,0 +1,1 @@
+dont_create_value = "(none)"